nosecurity+settings commit3

+++ PasswordGenerator.py
@@ -1,33 +1,39 @@
 
-#1h 00m
+#2h 41m
 
 import time
 import subprocess
 import sys
+import os
+from os import path
 
-
-def entrance():
-    try:
+def access(choice):                                                     # 0 -> prints list of userPasswords, 1 -> changes pin
+    if path.exists("savedPin.txt"):
         savedPin = open('savedPin.txt', 'r').read()
         pin = ""
         counter = 3
         while pin != savedPin:
-            pin = input("Insert your 4-digit pin to access the app: ")
-            if pin == savedPin:                                        #access granted
+            pin = input("Insert your 4-digit pin to access: ")
+            if pin == savedPin:                                         #access granted
                 print("You have access")
+                time.sleep(1)
+                if choice == 0:
+                    userPasswords(1)
+                elif choice == 1:
+                    changePin()
             else:
-                counter -=1
+                counter -= 1
                 if counter == 0:
                     print("Sorry the program will close now!")
                     time.sleep(1)
                     sys.exit()
                 print("Wrong Pin you have", counter, "guesses left")   #access denied
-    except IOError:
+    else:
         f = open("savedPin.txt", "w")                                  #if savedPin.txt does not exist, creates it
-        newPin = input("Please insert a new 4-digit pin: ")
+        newPin = input("Please insert a new 4-digit pin: ")             
         f.write(newPin)
         f.close()
-        subprocess.check_call(["attrib","+H","savedPin.txt"])          #converts the text file to hidden text file (my only security)
+        time.sleep(1)
 
 def mainMenu():
     print("Menu")
@@ -39,23 +45,72 @@ def mainMenu():
     if menuInput<5 and menuInput>0:
         if menuInput == 1:                                          #password generator
             generatePassword()
-        elif menuInput == 2:                                        #open entrance
-            entrance()
+            mainMenu()
+        elif menuInput == 2:                                        #open access
+            access(0)
+            mainMenu()
         elif menuInput == 3:                                        #open settings
-            print(2)
+            settings()
+            mainMenu()
         elif menuInput == 4:                                        #exit
-            sys.exit()
+            sys.exit()        
     else:
-        print("Sorry wrong input! Please try again.")
-        time.sleep(1)
+        print("Choose one of the following options: | '1' | '2' | '3' | '4' |")
+        print("Please try again.")
+        time.sleep(2)
         mainMenu()
     
 def generatePassword():
-    
+    #to be completed
+    return 0
 
+def pinCheck(pin):
+    #to be completed
+    return 0
+
+def settings():
+    print("To change your current pin press 1")
+    print("-Maybe a different format for the gened-password- press 2")
+    print("To go back to main menu press 3")
+    choiceInput = int(input())
+    if choiceInput == 1:                                                    #Changes pin
+        access(1)
+        f = open("savedPin.txt", "r")
+        for i in f:
+            print("Your new pin is: ", i)
+        f.close()
+        time.sleep(1)
+    elif choiceInput == 2:
+        print("-We will see-")
+        settings()
+    elif choiceInput == 3:
+        return None
+    else:
+        print("Choose one of the following options: | '1' | '2' | '3' |")
+        time.sleep(1)
+        settings()
+
+def userPasswords(choice):
+    try:
+        f = open("passwords.txt", "r")
+    except IOError:
+        f = open("passwords.txt", "w")
+        f.close
+    if choice == 1:                                                       #Ektypwsi tou arxeiou me ta passwords mono gia choice 1
+        for line in f:
+            print(line)
+        print
+    f.close()
+
+def changePin():
+    newPin = input("Insert here your new 4-digit pin: ")
+    pinfile = open("savedPin.txt", "w+")
+    pinfile.write(newPin)    
+    pinfile.close()
     
 
 if __name__ == "__main__":
+    userPasswords(0)
     print("_______WELCOME TO PASSWORD GENERATOR_______")
     time.sleep(1)
     mainMenu()
